requirements.txt
# Core dependencies
networkx>=2.8.0
pandas>=1.3.0
numpy>=1.21.0

# Language model integration
langchain-openai>=0.1.0
langchain-community>=0.2.0
python-dotenv>=1.0.0

# MCP (Model Context Protocol) for Serena integration
mcp>=1.0.0

# Community detection algorithms
python-louvain>=0.16
leidenalg>=0.9.0

# Visualization dependencies removed - text-only reports

# Code analysis
ast-comments>=1.0.0

# Utilities
pyyaml>=5.4.0
requests>=2.26.0
python-dateutil>=2.8.0
tqdm>=4.62.0

# Development and testing
pytest>=7.0.0
pytest-cov>=3.0.0
black>=22.0.0